+++ Sources/TuistGenerator/Linter/TargetActionLinter.swift
@@ -14,17 +14,30 @@ protocol TargetActionLinting {
 class TargetActionLinter: TargetActionLinting {
     func lint(_ action: TargetAction) -> [LintingIssue] {
         var issues: [LintingIssue] = []
+        issues.append(contentsOf: lintEmbeddedScriptNotEmpty(action))
         issues.append(contentsOf: lintToolExistence(action))
         issues.append(contentsOf: lintPathExistence(action))
         return issues
+    }
+
+    private func lintEmbeddedScriptNotEmpty(_ action: TargetAction) -> [LintingIssue] {
+        guard let script = action.embeddedScript,
+            script.isEmpty
+        else { return [] }
+
+        return [
+            LintingIssue(reason: "The embedded script is empty", severity: .warning),
+        ]
     }
 
     /// Lints a target aciton.
     ///
     /// - Parameter action: Action to be linted.
     /// - Returns: Found linting issues.
-    func lintToolExistence(_ action: TargetAction) -> [LintingIssue] {
-        guard let tool = action.tool else { return [] }
+    private func lintToolExistence(_ action: TargetAction) -> [LintingIssue] {
+        guard
+            let tool = action.tool
+        else { return [] }
         do {
             _ = try System.shared.which(tool)
             return []
@@ -34,7 +47,7 @@ class TargetActionLinter: TargetActionLinting {
         }
     }
 
-    func lintPathExistence(_ action: TargetAction) -> [LintingIssue] {
+    private func lintPathExistence(_ action: TargetAction) -> [LintingIssue] {
         guard
             let path = action.path,
             !FileHandler.shared.exists(path)