+++ Sources/TuistLoader/Models+ManifestMappers/TargetAction+ManifestMapper.swift
@@ -11,20 +11,29 @@ extension TuistCore.TargetAction {
     ///   - generatorPaths: Generator paths.
     static func from(manifest: ProjectDescription.TargetAction, generatorPaths: GeneratorPaths) throws -> TuistCore.TargetAction {
         let name = manifest.name
-        let tool = manifest.tool
         let order = TuistCore.TargetAction.Order.from(manifest: manifest.order)
-        let arguments = manifest.arguments
         let inputPaths = try manifest.inputPaths.map { try generatorPaths.resolve(path: $0) }
         let inputFileListPaths = try manifest.inputFileListPaths.map { try generatorPaths.resolve(path: $0) }
         let outputPaths = try manifest.outputPaths.map { try generatorPaths.resolve(path: $0) }
         let outputFileListPaths = try manifest.outputFileListPaths.map { try generatorPaths.resolve(path: $0) }
         let basedOnDependencyAnalysis = manifest.basedOnDependencyAnalysis
-        let path = try manifest.path.map { try generatorPaths.resolve(path: $0) }
+
+        let script: TuistCore.TargetAction.Script
+        switch manifest.script {
+        case let .embedded(text):
+            script = .embedded(text)
+
+        case let .scriptPath(path, arguments):
+            let scriptPath = try generatorPaths.resolve(path: path)
+            script = .scriptPath(scriptPath, args: arguments)
+
+        case let .tool(tool, arguments):
+            script = .tool(tool, arguments)
+        }
+
         return TargetAction(name: name,
                             order: order,
-                            tool: tool,
-                            path: path,
-                            arguments: arguments,
+                            script: script,
                             inputPaths: inputPaths,
                             inputFileListPaths: inputFileListPaths,
                             outputPaths: outputPaths,